core/project/settings/templates/settings.dev.py
import os
from pathlib import Path

from dotenv import load_dotenv

dotenv_path = Path.joinpath(Path(__file__).parent.parent.parent.parent.resolve(), '.dev.env')
load_dotenv(dotenv_path)

DEBUG = True
SECRET_KEY = 'django-insecure-=okvqjswarmd2ymc#jxqo-t7a(9qpvk-dk8v#9$t^c(dlbmw)b'
STATIC_ROOT = '/opt/project/static/'
MEDIA_ROOT = './media/'

CACHES = {
    'default': {
        'BACKEND': 'django_redis.cache.RedisCache',
        'LOCATION': 'redis://localhost:6379/0',
        'OPTIONS': {'CLIENT_CLASS': 'django_redis.client.DefaultClient'},
    },
}

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql',
        'NAME': os.environ.get('POSTGRES_DB'),
        'USER': os.environ.get('POSTGRES_USER'),
        'PASSWORD': os.environ.get('POSTGRES_PASSWORD'),
        'HOST': 'localhost',
        'PORT': '5432',
        'ATOMIC_REQUESTS': True,
        'CONN_MAX_AGE': 600,
    }
}
