pyproject.toml
[tool.poetry]
name = "affiliate-backend"
version = "0.1.0"
description = ""
authors = ["scxrchy <scxrchybeats@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.4.10"
django = "^5.0.6"
django-split-settings = "^1.3.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
django-redis = "^5.4.0"
django-cors-headers = "^4.4.0"
gunicorn = "^22.0.0"
pre-commit = "^3.7.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"
django-jazzmin = "^3.0.0"
daphne = "^4.1.2"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.6.0"
redis = "^5.0.6"
django-celery-results = "^2.5.1"
requests = "^2.32.3"
retrying = "^1.3.4"
ijson = "^3.3.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
